guokr-cmdb/requirements.txt
alembic==0.9.8
amqp==2.2.2
aniso8601==2.0.0
ansible==2.5.1
asn1crypto==0.24.0
astroid==1.6.3
awscli==1.14.22
Babel==2.5.3
bcrypt==3.1.4
billiard==3.5.0.3
boto3==1.5.12
botocore==1.8.26
celery==4.1.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.2.2
decorator==4.2.1
Django==2.0
docutils==0.14
dpath==1.4.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flower==0.9.2
get==0.0.33
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
influxdb==5.0.0
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kombu==4.1.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
paramiko==2.4.1
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
public==0.0.59
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pylint==1.8.4
PyMySQL==0.8.0
PyNaCl==1.2.1
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
query-string==0.0.21
redis==2.10.6
request==0.0.22
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.0
swagger-py-codegen==0.2.9
tornado==5.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
win-unicode-console==0.5
wrapt==1.10.11
PyJWT==1.6.3
py-zabbix==1.1.3
